more CleanUp

+++ TagCloud/Parsers/WordsToTagsParser.cs
@@ -19,15 +19,15 @@ namespace TagCloud
         public Result<List<Tag>> GetTags(Dictionary<string, int> words, ImageSettings imageSettings)
         {
             return Result.Of(() => fontSettingsFactory.CreateFontSettingsOrThrow(imageSettings.FontName))
-                .Then(fs => words.Select(s => new Tag(s.Key, s.Value, GetFont(fs, s.Value,words)))
+                .Then(fs => words.Select(s => new Tag(s.Key, s.Value, GetFont(fs, s.Value, words)))
                     .OrderByDescending(t => t.Count)
                     .ToList());
         }
 
-        private Font GetFont(FontSettings fontSettings, int count, Dictionary<string,int> words)
+        private Font GetFont(FontSettings fontSettings, int count, Dictionary<string, int> words)
         {
-            var fontSize= fontSettings.defaultFontSize + count * 300 / words.Count ;
-            return new Font(fontSettings.fontFamilyName, (float) fontSize, fontSettings.fontStyle);
+            var fontSize = fontSettings.defaultFontSize + count * 300 / words.Count;
+            return new Font(fontSettings.fontFamilyName, fontSize, fontSettings.fontStyle);
         }
     }
 }
